question_AND_answering_new/haggingfacedir/requirements.txt
pillow
transformers
requests